+++ packages/adapter-nextjs/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.5.8...@aws-amplify/adapter-nextjs@1.6.0) (2025-04-09)
+
+### Features
+
+- **adapter-nextjs:** support sending lang param to managed login ([#14322](https://github.com/aws-amplify/amplify-js/issues/14322)) ([d11c5ce](https://github.com/aws-amplify/amplify-js/commit/d11c5ce413219cc88b7b68fb5b2c2c9050a5c6b1))
+
+## [1.5.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.5.7...@aws-amplify/adapter-nextjs@1.5.8) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/adapter-nextjs
+
 ## [1.5.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/adapter-nextjs@1.5.6...@aws-amplify/adapter-nextjs@1.5.7) (2025-03-25)
 
 **Note:** Version bump only for package @aws-amplify/adapter-nextjs