+++ packages/api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@6.3.8...@aws-amplify/api@6.3.9) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## [6.3.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@6.3.7...@aws-amplify/api@6.3.8) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/api
+
 ## [6.3.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@6.3.6...@aws-amplify/api@6.3.7) (2025-03-25)
 
 **Note:** Version bump only for package @aws-amplify/api