+++ packages/auth/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.12.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.12.1...@aws-amplify/auth@6.12.2) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/auth
+
+## [6.12.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.12.0...@aws-amplify/auth@6.12.1) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/auth
+
 # [6.12.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.11.6...@aws-amplify/auth@6.12.0) (2025-03-25)
 
 ### Features