+++ packages/datastore-storage-adapter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.80](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.1.79...@aws-amplify/datastore-storage-adapter@2.1.80) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
+## [2.1.79](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.1.78...@aws-amplify/datastore-storage-adapter@2.1.79) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
 ## [2.1.78](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.1.77...@aws-amplify/datastore-storage-adapter@2.1.78) (2025-03-25)
 
 **Note:** Version bump only for package @aws-amplify/datastore-storage-adapter