+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.78](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.77...@aws-amplify/geo@3.0.78) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.77](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.76...@aws-amplify/geo@3.0.77) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## [3.0.76](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.75...@aws-amplify/geo@3.0.76) (2025-03-25)
 
 **Note:** Version bump only for package @aws-amplify/geo