+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.53](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.52...@aws-amplify/pubsub@6.1.53) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.52](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.51...@aws-amplify/pubsub@6.1.52) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
 ## [6.1.51](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.50...@aws-amplify/pubsub@6.1.51) (2025-03-25)
 
 **Note:** Version bump only for package @aws-amplify/pubsub