+++ packages/storage/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.8.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.8.1...@aws-amplify/storage@6.8.2) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
+## [6.8.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.8.0...@aws-amplify/storage@6.8.1) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
 # [6.8.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@6.7.16...@aws-amplify/storage@6.8.0) (2025-03-25)
 
 ### Features