+++ scripts/tsc-compliance-test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.83](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.82...tsc-compliance-test@0.1.83) (2025-04-09)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.82](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.81...tsc-compliance-test@0.1.82) (2025-03-28)
+
+**Note:** Version bump only for package tsc-compliance-test
+
 ## [0.1.81](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.80...tsc-compliance-test@0.1.81) (2025-03-25)
 
 **Note:** Version bump only for package tsc-compliance-test